requirements.txt
fastapi
uvicorn
opencv-python-headless
numpy
torch
torchvision
pillow
jinja2
git+https://github.com/xinntao/BasicSR.git
